Generate fake database records
Generate fake records , Model factory and seeder

Model factory which generates the fake data and seeding which loads the fake data into the database.

Naming a factory class : start with the model name followed by Factory, e.g taskFactory this will enable laravel to automatically tell for which model that class will be factory will be used for.

fake() is a library function associated with laravel which is responsible for loading the fake data and method fake is used in the factory for the

To create a new factory model we use  the command "php artisan make:factory TaskFactory --model=Task"

Roll back database seed migrations "php artisan migrate:refresh --seed" Never run the migrate:refresh command on the live object as it will wipe out the database