Forms and CSRF protection
CRF tokens

typical name of blade template responsible for creating fors is create.blade.php

In case you don't need to fetch additional data in you route you can use the view method instead of get mrthod

-The  dd()  function stands for "Dump and Die". It will display the provided value or variable and immediately terminate the script execution. This function is commonly used to inspect the value of variables or to check if a specific code block is being executed.

-@csrf :: @csrf is a Blade directive in Laravel that is used to generate a hidden input field containing a CSRF token. This token is used to verify that the authenticated user is the one actually making the requests to the application.
Sure. @csrf is a Blade directive in Laravel that is used to generate a hidden input field containing a CSRF token. This token is used to verify that the authenticated user is the one actually making the requests to the application.

CSRF stands for Cross-Site Request Forgery. It is a type of malicious exploit whereby unauthorized commands are performed on behalf of an authenticated user.