validation
validation

{{ $errors }} ::
The {{ $errors }} Blade directive in Laravel is used to display error messages for form fields. The directive will render a list of all of the error messages for the current form.

The {{ $errors }} directive is typically used in conjunction with the @if directive. For example, the following code would display error messages for the name and email form fields if there are any errors: